+++ CHANGELOG.md
@@ -4,3 +4,5 @@ Change Log
 0.0.1 (April 9th, 2025)
 -----------------------
 - First Release
+- Added the course and courseSection classes
+- Added the getCourseContentByID method